backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
alembic>=1.13.0
python-multipart>=0.0.6
email-validator>=2.0.0
httpx>=0.25.0
pytest>=7.4.0
